Finish refactoring report server code.
